requirements.txt
streamlit==1.39.0
openai==1.52.1
toml~=0.10.2
pandas==2.2.3
tabulate==0.9.0
python-dotenv==1.0.1
streamlit_ace==0.1.1

PyPDF2==3.0.1
pdf2image==1.17.0
pillow==10.4.0
scikit-learn==1.5.2
tiktoken==0.8.0
numpy==2.1.2

pyaudio==0.2.14
gtts==2.5.3
pygame==2.6.1

urllib3==2.2.3
httpx==0.27.2
